added game.html and script.js to handle user difficulty choice
