+++ game.html
@@ -1,0 +1,52 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+    <title>Word Scramble</title>
+    <link rel="stylesheet" href="./assets/style1.css" />
+    <script defer src="./js/script_easy.js" type="text/javascript"></script>
+    <script defer src="./js/script_medium.js" type="text/javascript"></script>
+    <script defer src="./js/script_hard.js" type="text/javascript"></script>
+</head>
+
+<body>
+    <header>
+        <h1>Welcome to Word Scramble!</h1>
+        <nav>
+            <button class="start-button">Start</button>
+        </nav>
+    </header>
+    <main>
+        <section>
+            <div class="card word-guess">
+                <div class="large-font word-blanks">R-a-m-b-l-e-S-c</div>
+            </div>
+        </section>
+        <section>
+            <div class="card results">
+                <div class="win-loss-container">
+                    <div>
+                        <h2>Wins: <span class="win">0 </span> </h2>
+                        <h2>Losses: <span class="lose">0</span></h2>
+                    </div>
+                    <button class="reset-button">Reset Score</button>
+                </div>
+                <div class="card timer">
+                    <div class="timer-text">
+                        <div class="large-font timer-count">60</div>
+                        <h3>seconds remaining</h3>
+                    </div>
+
+                </div>
+            </div>
+        </section>
+    </main>
+    <footer>
+        <p id="love">Made by Team 1!</p>
+    </footer>
+</body>
+
+</html>
