+++ index.html
@@ -6,8 +6,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-    <script src="script.js"></script>
 
+    <link rel="stylesheet" type="text/css" href="./assets/style_easy.css" />
 </head>
 
 <html class="h-full bg-white">
@@ -16,18 +16,19 @@
     <div class="d-flex flex-column align-items-center">
         <!-- Image container -->
         <div class="image-container">
-            <img src="./assets/image1.jpeg" class="img-fluid mx-auto d-block p-3" alt="Word Scramble pic">
-            <h1>"Welcome to Word Scramble!</h1> 
+            <img src="./assets/image1.jpeg" class="w-100 img-fluid mx-auto d-block p-3" alt="Word Scramble pic">
+            <h1>"Welcome to Word Scramble!</h1>
         </div>
-        
+
         <!-- Button container -->
         <div class="button-container d-flex gap-3"> <!-- gap-3 adds space between buttons -->
             <button type="button" class="btn btn-primary">
                 Play Game!
             </button>
-        
+
             <div class="dropdown">
-                <button class="btn btn-secondary dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">
+                <button class="btn btn-secondary dropdown-toggle" type="button" id="difficultyButton" data-bs-toggle="dropdown"
+                    aria-expanded="false">
                     Choose difficulty!
                 </button>
                 <ul class="dropdown-menu">
@@ -38,7 +39,7 @@
             </div>
         </div>
     </div>
-
+    <script src="script.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
         crossorigin="anonymous"></script>
